backend/requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
boto3==1.21.8
botocore==1.24.8
dj-database-url==0.5.0
Django==4.0.2
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.2
importlib-metadata==4.11.1
importlib-resources==5.4.0
inflection==0.5.1
jmespath==0.10.0
jsonschema==4.4.0
Markdown==3.3.6
Pillow==9.0.1
psycopg2==2.9.3
PyJWT==2.3.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
